+++ R/zzz.R
@@ -5,7 +5,7 @@ defaultPkgOpts <- function(pkgname)
                 maxProcAmount = parallel::detectCores(),
                 path.pwiz = "",
                 path.GenForm = "",
-                path.metFragCL = "",
+                path.MetFragCL = getOption("patRoon.path.metFragCL", ""), # backwards compat
                 path.SIRIUS = "",
                 path.OpenMS = "",
                 path.pngquant = "")